Reorganize expected test results for better overview and reuse
- move expected result json files to a `data` folder
- eliminate duplication of json loading helpers
