+++ tests/data/aggregate_zonal.json
