+++ tests/data/apply_dimension_temporal_cumsum.json
