+++ tests/test_bandmath.py
@@ -1,23 +1,11 @@
-import os
 import unittest
 from unittest import TestCase
 
-from mock import MagicMock
 import requests_mock
-from pathlib import Path
+from mock import MagicMock
 
 import openeo
-
-
-def get_test_resource(relative_path):
-    dir = Path(os.path.dirname(os.path.realpath(__file__)))
-    return str(dir / relative_path)
-
-
-def load_json_resource(relative_path):
-    import json
-    with open(get_test_resource(relative_path), 'r+') as f:
-        return json.load(f)
+from tests import load_json_resource
 
 
 @requests_mock.mock()
@@ -56,7 +44,7 @@ class TestBandMath(TestCase):
 
         session.download.assert_called_once()
         actual_graph = session.download.call_args_list[0][0][0]
-        expected_graph = load_json_resource('evi_graph.json')
+        expected_graph = load_json_resource('data/evi_graph.json')
         assert actual_graph == expected_graph
 
     def test_ndvi(self, m):
@@ -176,7 +164,7 @@ class TestBandMath(TestCase):
                                                                                         'to': '2018-06-18'}})
 
         def check_process_graph(request):
-            expected_graph = load_json_resource('udf_graph.json')
+            expected_graph = load_json_resource('data/udf_graph.json')
             assert request.json() == expected_graph
             return True
